+++ CHANGELOG.md
@@ -14,6 +14,7 @@
  - Prevent pre-output from remaining on after an Input is deactivated
  - Enable unit conversions for AM2315
  - Fix issue setting PID setpoint from Dashboard ([#496](https://github.com/kizniche/mycodo/issues/496))
+ - Fix displaying custom graph colors ([#491](https://github.com/kizniche/mycodo/issues/491))
 
 ### Miscellaneous